% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_qwi_labels.R
\name{add_qwi_labels}
\alias{add_qwi_labels}
\title{add_qwi_labels}
\usage{
add_qwi_labels(df)
}
\arguments{
\item{df}{an object with a class of `qwi`}
}
\description{
This function add labels to a `qwi` object
}
\examples{
\donttest{
library(tidyqwi)
labelled_nc <- add_qwi_labels(nc_qwi)

}

}
