% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fips.R
\name{converted_fips}
\alias{converted_fips}
\title{A function to check if a valid state number or fips is passed}
\usage{
converted_fips(fips)
}
\arguments{
\item{fips}{the state abbreviation or fips code vector}
}
\value{
States Abbreviations or FIPs as FIP character strings
}
\description{
A function to check if a valid state number or fips is passed
}
