% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_codes.R
\docType{data}
\name{geo_codes}
\alias{geo_codes}
\title{Geographic Codes}
\format{
a dataframe with 5423 rows and 3 columns:
\describe{
  \item{geography}{geography}
  \item{label}{label}
  \item{geolevel}{geolevel}
   ...
}
}
\source{
\url{https://lehd.ces.census.gov/data/schema/latest/label_geography.csv}
}
\usage{
geo_codes
}
\description{
Geographic Codes
}
\keyword{datasets}
