% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/industry_labels.R
\docType{data}
\name{industry_labels}
\alias{industry_labels}
\title{Industry Labels}
\format{a dataframe with 433 rows and 3 columns:
\describe{
  \item{industry}{Industry Numeric Code}
  \item{label}{Description of Industry Level}
  \item{ind_level} {Industry Level}
   ...
}}
\source{
\url{https://lehd.ces.census.gov/data/schema/latest/label_industry.csv}
}
\usage{
industry_labels
}
\description{
Industry Labels
}
\keyword{datasets}
