% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_qwi.R
\docType{data}
\name{nc_qwi}
\alias{nc_qwi}
\title{QWI Variable Names}
\format{a dataframe with 3244 rows and 44 columns:}
\source{
\url{https://api.census.gov/data/timeseries/qwi/se/variables.html}
}
\usage{
nc_qwi
}
\description{
These data represent an example returned query for NC for 2010
}
\keyword{datasets}
