% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owner_codes.R
\docType{data}
\name{owner_codes}
\alias{owner_codes}
\title{Owner  Codes}
\format{a dataframe with 3 rows and 2 columns:
\describe{
  \item{ownercode}{ownercode}
  \item{label}{label}
   ...
}}
\source{
\url{https://lehd.ces.census.gov/data/schema/latest/label_ownercode.csv}
}
\usage{
owner_codes
}
\description{
Owner  Codes
}
\keyword{datasets}
