% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qwi_message.R
\name{parse_qwi_message}
\alias{parse_qwi_message}
\title{parse_qwi}
\usage{
parse_qwi_message(x)
}
\arguments{
\item{x}{a returned call response from the US Census QWI API}
}
\description{
parse_qwi
}
