% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_info.R
\docType{data}
\name{state_info}
\alias{state_info}
\title{State Data (FIPS, Abbreviations, etc)}
\format{a dataframe with 51 rows and 3 columns:
\describe{
  \item{name}{state name}
  \item{state_fips}{state fips code}
  \item{state_abbreviation} {state abreviation}
   ...
}}
\source{
\url{https://www.census.gov/geo/reference/ansi_statetables.html}
}
\usage{
state_info
}
\description{
State Data (FIPS, Abbreviations, etc)
}
\keyword{datasets}
