% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwi_var_names.R
\docType{data}
\name{qwi_var_names}
\alias{qwi_var_names}
\title{QWI Variable Names}
\format{a dataframe with 83 rows and 9 columns:
\describe{
  \item{name}{state name}
  \item{label}{state fips code}
  \item{concept} {state abreviation}
  \item{required}
  \item{attributes}
  \item{limit}
  \item{predicate type}
  \item{group}
  \item{values}
   ...
}}
\source{
\url{https://api.census.gov/data/timeseries/qwi/se/variables.html}
}
\usage{
qwi_var_names
}
\description{
QWI Variable Names
}
\keyword{datasets}
